+++ src/main/app/src/components/tables/CustomToolbar.js
@@ -1,0 +1,16 @@
+import {
+    GridToolbarContainer, GridToolbarExport, GridColumnsToolbarButton,
+    GridFilterToolbarButton
+  } from '@material-ui/data-grid';
+
+function CustomToolbar() {
+    return (
+      <GridToolbarContainer>
+        <GridColumnsToolbarButton />
+        <GridFilterToolbarButton />
+        <GridToolbarExport />
+      </GridToolbarContainer>
+    );
+  }
+
+export default CustomToolbar;
